+++ src/oms_sensemaking/clients/instances.py
@@ -60,7 +60,7 @@ def get_db_session() -> Iterator[Session]:
 
 
 aac_client = AacClient(SETTINGS.cert_path, SETTINGS.key_path, SETTINGS.cacert_path, SETTINGS.aac_verification_mode)
-oms_client = OmsCrudTool()
+
 oms_crud_tool = OmsCrudTool()
 
 corenlp_client = CoreNlpClient(props=SETTINGS.corenlp_client_props, hostname=SETTINGS.corenlp_host)